Break up single gigantic test into multiple tests
We also explicitly stringify stdout/stderr contents as this seems to
mitigate some flakiness. (Emphasis on seems to)

Note that because ginkgo doesn't have a BeforeAll and you can only run a
single BeforeSuite, we need to set up the API and login before each
test. This is a bit unfortunate but sometimes it takes a tough
programmer to make a tender chicken. Also, you can run the tests with
the -p flag to parallelize them and speed them up.

You're welcome. :)

[#130699549] User should be able to generate SSH key credential

Signed-off-by: Sebastian Wittenkamp <swittenkamp@pivotal.io>