Kludge over differences between CI and local dev
With identical setup and identical 1.7.5 go toolchains, the test behaves
differently in each environment. The core difference is in OS. Without
digging further, we have struck a compromise.

Given a cert signed by an unknown CA, the Linux version does not send the
certificate. The server errors with "Full authentication is required...".

Given a cert signed by an unknown CA, the OS X version errors, we're not
sure if the error is the server rejecting the cert or OS X refusing to
send it. In this case, Golang itself hands back an err, containing an
"unknown certificate" message.

[#138290503] Client cert validity dates should be validated for MTLS authentication

Signed-off-by: Jacques Chester <jchester@pivotal.io>