Lower risk level for SQLmap run
- Since the current job in CI seems to be running for durations longer
than 24 hours, I am dialling down the risk level and sqlmap level
temporarily to validate that we can get a green run on CI as we haven't
had one yet. Once we get a green build, we can revisit our plan of
action regarding the run frequency and configuration of the same.

[#142155925] Investigate using SQLMap tool against CredHub to identify any SQLi