feat: augment TLS compatibility with backends
- full context in: https://github.com/pivotal/credhub-release/commit/b6a6fac08372abc770864785619e2335d914677d
- TLDR: - Before, CredHub specifies that "secp384r1"
elliptic curve be used for TLS connections. We found no clear
justifications for this specification (the commit and story
associated with this specification, #154184086, did not explain
why). We decide to remove this specification, and instead,
rely on Java's default elliptic curve name groups,
which is an ordered list of elliptic curve groups deemed
acceptable by Java. Hence, this commit adjusts the tests
to no longer assert on which specific elliptic curve groups
are used.

[#183852496]

Co-authored-by: Bruce Ricard <bricard@vmware.com>